requirements.txt
amqp==5.2.0
asgiref==3.7.2
astroid==3.2.2
billiard==4.2.0
black==24.4.2
celery==5.4.0
certifi==2023.11.17
cffi==1.16.0
channels==4.1.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
contourpy==1.2.0
cron-descriptor==1.4.3
cryptography==42.0.8
cycler==0.12.1
defusedxml==0.7.1
diff-match-patch==20230430
dill==0.3.8
Django==4.2.6
django-celery-beat==2.6.0
django-filter==23.4
django-import-export==3.3.3
django-redis==5.4.0
django-request==1.6.3
django-requests==0.0.2
django-simple-history==3.4.0
django-timezone-field==6.1.0
djangorestframework==3.14.0
drf-yasg==1.21.7
et-xmlfile==1.1.0
fonttools==4.49.0
idna==3.6
inflection==0.5.1
isort==5.13.2
kiwisolver==1.4.5
kombu==5.3.7
MarkupPy==1.14
matplotlib==3.8.3
mccabe==0.7.0
mypy-extensions==1.0.0
numpy==1.26.4
oauthlib==3.2.2
odfpy==1.4.1
openpyxl==3.1.2
packaging==23.2
pathspec==0.12.1
Pillow==10.1.0
platformdirs==4.2.2
prompt_toolkit==3.0.47
pycparser==2.22
PyJWT==2.8.0
pylint==3.2.3
pylint-django==2.5.5
pylint-plugin-utils==0.8.2
pyparsing==3.1.1
python-crontab==3.1.0
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2023.3.post1
PyYAML==6.0.1
redis==5.0.6
requests==2.31.0
requests-oauthlib==2.0.0
setuptools==70.1.0
six==1.16.0
social-auth-app-django==5.4.1
social-auth-core==4.5.4
sqlparse==0.4.4
tablib==3.5.0
tomlkit==0.12.5
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.1.0
vine==5.1.0
wcwidth==0.2.13
xlrd==2.0.1
xlwt==1.3.0
django-celery-results
flower
